+++ functions/Resume-MyTimer.ps1
@@ -15,8 +15,9 @@ Function Resume-MyTimer {
     )
 
     Begin {
-        Write-Verbose "[$((Get-Date).TimeOfDay) BEGIN  ] Starting $($MyInvocation.MyCommand)"
-        Write-Verbose "[$((Get-Date).TimeOfDay) BEGIN  ] Running under PowerShell version $($PSVersionTable.PSVersion)"
+        _verbose  ($strings.starting -f $MyInvocation.MyCommand)
+        _verbose ($strings.Running -f $PSVersionTable.PSVersion)
+        _verbose ($strings.Detected -f $host.Name)
     } #begin
 
     Process {
@@ -32,12 +33,12 @@ Function Resume-MyTimer {
         else {
             Write-Warning "You can only resume a paused timer. The timer '$($timer.name)' has a status of $($timer.Status)."
         }
-        Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] Pausing timer $Name"
+        _verbose "[$((Get-Date).TimeOfDay) PROCESS] Pausing timer $Name"
 
     } #process
 
     End {
-        Write-Verbose "[$((Get-Date).TimeOfDay) END    ] Ending $($MyInvocation.MyCommand)"
+        _verbose  ($strings.Ending -f  $MyInvocation.MyCommand)
     } #end
 
 } #close Resume-MyTimer