+++ functions/Stop-PSCountdownTimer.ps1
@@ -3,29 +3,30 @@ Function Stop-PSCountdownTimer {
     [OutputType("none")]
     Param( )
     Begin {
-        Write-Verbose "[$((Get-Date).TimeOfDay) BEGIN  ] Starting $($MyInvocation.MyCommand)"
+        _verbose  ($strings.starting -f $MyInvocation.MyCommand)
+        _verbose ($strings.Running -f $PSVersionTable.PSVersion)
+        _verbose ($strings.Detected -f $host.Name)
     } #begin
 
     Process {
         Try {
             [void](Get-Variable -Name PSCountdownClock -Scope global -ErrorAction Stop)
-            If ($PSCmdlet.ShouldProcess("countdown timer started at $($global:PSCountdownClock.started)")) {
+            If ($PSCmdlet.ShouldProcess($($Strings.StartedCountdown -f $global:PSCountdownClock.started))) {
                 $rs = $global:PSCountdownClock.RunSpace
                 $global:PSCountdownClock.running = $False
                 Remove-Variable -Name PSCountdownClock -Scope Global
                 Start-Sleep -Seconds 1
                 Remove-RunSpace $rs
             }
-
         }
         Catch {
-            Write-Warning "Could not find `$PSCountdownClock in the global scope. Did you start a countdown timer?"
+            Write-Warning $strings.NoCountDownTimer
         }
-        Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] Stopping PSCountdownClock"
-    } #process
+        _verbose ($strings.Ending -f $MyInvocation.MyCommand)
+    } #process"
 
     End {
-        Write-Verbose "[$((Get-Date).TimeOfDay) END    ] Ending $($MyInvocation.MyCommand)"
+        _verbose  ($strings.Ending -f  $MyInvocation.MyCommand)
     } #end
 
 } #close Stop-PSCountdownTimer