+++ Changelog.md
@@ -2,6 +2,21 @@
 
 ## [Unreleased]
 
+## [2.3.0] - 2025-08-21
+
+### Added
+
+- Added `Start-PSCountdownTitle` and its alias `TitleCountdown`. This command lets you run a countdown timer in the console title or tab, if using Windows Terminal.
+
+### Changed
+
+- Updated verbose messaging to provide runtime metadata.
+- Updated `Start-PSCountdownTimer` to run in VS Code.
+- Updated `README.md`.
+- Revised online help links and refreshed help files.
+- Added `Microsoft.Powershell.ThreadJob` as a dependency.
+- Moved primary git branch from `master` to `main`.
+
 ## [2.2.0] - 2024-09-05
 
 ### Changed
@@ -14,7 +29,7 @@
 
 - Fixed broken online help links.
 
-## [v2.1.0] - 2023-07-10
+## [2.1.0] - 2023-07-10
 
 ### Changed
 
@@ -25,7 +40,7 @@
 - Fixed default path reference in `Start-PSCountDown`.
 - Fixed `about_PSTimers` help topic.
 
-## [v2.0.1] - 2023-06-29
+## [2.0.1] - 2023-06-29
 
 ### Changed
 
@@ -33,7 +48,8 @@
 - Updated external help.
   Updated `README.md`.
 
-[Unreleased]: https://github.com/jdhitsolutions/PSTimers/compare/v2.2.0..HEAD
+[Unreleased]: https://github.com/jdhitsolutions/PSTimers/compare/v2.3.0..HEAD
+[2.3.0]: https://github.com/jdhitsolutions/PSTimers/compare/v2.2.0..v2.3.0
 [2.2.0]: https://github.com/jdhitsolutions/PSTimers/compare/v2.1.0..v2.2.0
-[v2.1.0]: https://github.com/jdhitsolutions/pstimers/compare/v2.0.1..v2.1.0
-[v2.0.1]: https://github.com/jdhitsolutions/pstimers/compare/v2.0.0..v2.0.1
+[2.1.0]: https://github.com/jdhitsolutions/pstimers/compare/v2.0.1..v2.1.0
+[2.0.1]: https://github.com/jdhitsolutions/pstimers/compare/v2.0.0..v2.0.1